# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: GraphSign REST API v1
  version: 1.0.0
  description: Official REST APIs for the graphsign.ink electronic signature platform (FR-016).
servers:
  - url: https://dev-graphsign-api.kunal-f9f.workers.dev
    description: Development API
  - url: http://localhost:8787
    description: Local Development Server

paths:
  /api/v1/documents:
    get:
      summary: List documents
      description: Retrieve paginated documents with status, folder, and date range filters.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SENT, IN_PROGRESS, COMPLETED, DECLINED, CANCELLED, EXPIRED]
        - name: folder
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A list of documents with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      hasMore:
                        type: boolean
                      nextCursor:
                        type: string
                        nullable: true
    post:
      summary: Create a document
      description: Creates a new agreement document from text/markdown or base64 PDF.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentInput'
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '200':
          description: Idempotent replay of previously created document
        '400':
          description: Invalid request payload
        '409':
          description: Idempotency conflict

  /api/v1/documents/{id}:
    get:
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
    put:
      summary: Replace document
      description: Replace editable draft representation.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceDocumentInput'
      responses:
        '200':
          description: Document updated
    patch:
      summary: Partially update document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDocumentInput'
      responses:
        '200':
          description: Document updated
    delete:
      summary: Soft-delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document successfully deleted

  /api/v1/webhooks:
    get:
      summary: List webhook subscriptions
      responses:
        '200':
          description: List of active webhook subscriptions
    post:
      summary: Create webhook subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookInput'
      responses:
        '201':
          description: Webhook subscription created with one-time secret

  /api/v1/health:
    get:
      summary: Public platform readiness check
      responses:
        '200':
          description: Platform dependencies are healthy
        '503':
          description: Platform dependencies unavailable

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Document:
      type: object
      required:
        - id
        - name
        - status
        - version
        - mimeType
        - fileSize
        - tags
        - metadata
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [DRAFT, SENT, IN_PROGRESS, COMPLETED, DECLINED, CANCELLED, EXPIRED]
        version:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateDocumentInput:
      type: object
      required:
        - name
        - content
      properties:
        name:
          type: string
        content:
          type: string
        contentEncoding:
          type: string
          enum: [utf-8, base64]
          default: utf-8
        mimeType:
          type: string
          enum: [application/pdf, text/markdown, text/plain]
          default: application/pdf
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          properties:
            folder:
              type: string

    ReplaceDocumentInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        content:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object

    PatchDocumentInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object

    CreateWebhookInput:
      type: object
      required:
        - name
        - targetUrl
        - eventTypes
      properties:
        name:
          type: string
        description:
          type: string
        targetUrl:
          type: string
          format: uri
        eventTypes:
          type: array
          items:
            type: string
        rateLimitPerMinute:
          type: integer
          default: 10
        filterRules:
          type: object
        payloadProjection:
          type: object
          properties:
            mode:
              type: string
              enum: [ALL, CUSTOM]
            includeFields:
              type: array
              items:
                type: string

security:
  - BearerAuth: []